backend/requirements.txt
# Core dependencies
pandas>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

# API server dependencies
Flask>=3.0.0
flask-cors>=4.0.0
PyJWT>=2.8.0

# Database driver
pyodbc>=5.0.0

# Additional utilities
python-json-logger>=2.0.0
tabulate>=0.9.0

# Testing
pytest>=7.0.0
requests>=2.28.0
